Refactor controllers out for easier testing
